Update config from electron-react-boilderplate

+++ .erb/scripts/clean.js
@@ -1,8 +1,9 @@
+import fs from 'fs';
 import rimraf from 'rimraf';
 import webpackPaths from '../configs/webpack.paths';
 
 const foldersToRemove = [webpackPaths.distPath, webpackPaths.buildPath, webpackPaths.dllPath];
 
 foldersToRemove.forEach((folder) => {
-  rimraf.sync(folder);
+  if (fs.existsSync(folder)) rimraf.sync(folder);
 });
